Allow filtering of log messages before publishing
We want to be able to send a subset of all journald messages off to a
different Kafka topic to all other messages. In order to select these
messages we want to be able to filter messages based on values in a
field.

For example if we want only STDOUT from a container with a docker label
of application=nginxlb, we can do the following:

```
loglet --include-filter APPLICATION=nginxlb --exclude-filter PRIORITY=3
```

journald will capitalize all fields and removing anything with
`PRIORITY=3` is the equivalent of dropping STDERR.

At the moment this allows multiple filters that will be combined with OR
semantics and only does exact matching.